wip: split canonical query classifiers
