+++ .agents/scripts/canonical_git_readonly.py
@@ -7,6 +7,8 @@ from __future__ import annotations
 
 from typing import Callable
 
+from canonical_git_ref_queries import REF_QUERY_CHECKS
+
 
 def _branch_is_read_only(args: list[str]) -> bool:
     if not args:
@@ -78,36 +80,9 @@ def _clean_is_read_only(args: list[str]) -> bool:
     )
 
 
-def _remote_is_read_only(args: list[str]) -> bool:
-    return not args or args[0] in {"-v", "--verbose", "get-url", "show"}
-
-
-def _worktree_is_allowed(args: list[str]) -> bool:
-    return bool(args) and args[0] in {"list", "add"}
-
-
-def _tag_is_read_only(args: list[str]) -> bool:
-    return not args or any(
-        arg in {"-l", "--list", "--contains", "--points-at"}
-        or arg.startswith(("--contains=", "--points-at="))
-        for arg in args
-    )
-
-
-def _symbolic_ref_is_read_only(args: list[str]) -> bool:
-    read_flags = {"-q", "--quiet", "--short", "--recurse", "--no-recurse"}
-    return (
-        sum(arg not in read_flags for arg in args) == 1
-        and all(arg in read_flags or not arg.startswith("-") for arg in args)
-    )
-
-
 CANONICAL_CHECKS: dict[str, Callable[[list[str]], bool]] = {
     "branch": _branch_is_read_only,
     "config": _config_is_read_only,
     "clean": _clean_is_read_only,
-    "remote": _remote_is_read_only,
-    "symbolic-ref": _symbolic_ref_is_read_only,
-    "worktree": _worktree_is_allowed,
-    "tag": _tag_is_read_only,
+    **REF_QUERY_CHECKS,
 }
